% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-preds.R
\name{calibrate_preds}
\alias{calibrate_preds}
\title{Recalibrate probabilities after computing logit shifts}
\usage{
calibrate_preds(
  ps_table,
  shifts,
  preds,
  geography,
  shift_suffix = "shift",
  keep_orig = TRUE
)
}
\arguments{
\item{ps_table}{A data frame storing poststratification cells with (uncalibrated)
model-based predictions.}

\item{shifts}{A data frame of logit shift parameters computed by \code{\link[=logit_shift]{logit_shift()}}.
Each shift must be named \verb{\{pred\}_\{shift_suffix\}}, e.g., \code{voteshare_shift} if
the prediction variable in \code{ps_table} is \code{voteshare}.}

\item{preds}{One or more prediction variables in \code{ps_table} to recalibrate.
Accepts tidyselect syntax (e.g., \code{var}, \code{c(var1, var2)}, \code{starts_with()}).}

\item{geography}{Group variable variable linking \code{ps_table} and \code{shifts}.
This should be a single variable name (unquoted) that exists in both data frames.}

\item{shift_suffix}{String suffix for the shift variables, e.g. if the first outcome
variable is called \code{presvote} then the shift variable will be called
\code{presvote_shift}.}

\item{keep_orig}{keep original uncalibrated columns?}
}
\value{
Returns \code{ps_table} with additional columns for calibrated predictions.
}
\description{
Takes the poststratification table and a table of logit shift parameters and
returns the calibrated cell-level probabilities.
}
\examples{
## See examples in ?logit_shift()
}
