% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_estimates.R
\name{get_outcomes}
\alias{get_outcomes}
\title{Infer outcomes from a model formula}
\usage{
get_outcomes(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
character vector of outcome variable(s)
}
\description{
Infer outcomes from a model formula
}
