% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-shift.R
\name{get_re_covs}
\alias{get_re_covs}
\title{Extract covariance of random effects across outcomes}
\usage{
get_re_covs(mod, group, tidy = FALSE, outcome_order = NULL, draw_ids = NULL)
}
\arguments{
\item{mod}{A \code{brmsfit} object with random intercepts by \code{group}.}

\item{group}{Grouping variable to extract random intercept.}

\item{tidy}{Return a tidy data frame? If \code{FALSE}, returns an array of
correlation matrices where the first dimension indexes draws from the
posterior.}

\item{outcome_order}{An order for the outcome variables to return. Only used
if tidy = FALSE.}

\item{draw_ids}{An integer vector specifying the posterior draws to be used.
If NULL (the default), all draws are used.}
}
\description{
\code{get_re_covs} returns the covariance parameters for random effects that are
correlated across outcomes. It takes a \code{brmsfit} object and extracts the parameters
that start with \code{sd_} and \code{cor_}, end with \verb{_Intercept}, and include the pattern
given in the \code{group} argument. If \code{tidy = TRUE}, the output is a tidy data frame
in the format of \code{tidybayes::spread_draws}, except with columns given informative
names. If \code{tidy = FALSE}, an array of covariance matrices is returned (one
for each draw from the posterior).
}
