% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-shift.R
\name{logit_shift_aux}
\alias{logit_shift_aux}
\title{Estimate logit shift for auxiliary variables given calibrated outcomes and covariance}
\usage{
logit_shift_aux(shift, shift_vars, cov, suffix = "_shift")
}
\arguments{
\item{shift}{A data frame storing original logit shifts for each level of geography.}

\item{shift_vars}{Variables in \code{shift} data frame that stores the logit shifts.}

\item{cov}{A covariance matrix for the logit shift across outcomes. Row
and column names should contain the variables in \code{shift_vars}.}

\item{suffix}{The suffix for the logit shift parameters in \code{shift}. The
prefixes are given by \code{shift_vars}. For example if the outcome is \code{dem} then
the logit shift parameter might be stored in a variable called \code{dem_shift}.}
}
\description{
Estimate logit shift for auxiliary variables given calibrated outcomes and covariance
}
