% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-shift.R
\name{logit_shift_intercept}
\alias{logit_shift_intercept}
\title{Shift probabilities on the logit scale}
\usage{
logit_shift_intercept(x, a)
}
\description{
Shift probabilities on the logit scale
}
\keyword{internal}
