% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-shift.R
\name{logit_shift_internal}
\alias{logit_shift_internal}
\title{Find the logit-scale intercept shift such that weighted.mean(x,w) = target}
\usage{
logit_shift_internal(
  x,
  w = rep(1, length(x)),
  target,
  na.rm = TRUE,
  tol = 1e-05
)
}
\description{
Find the logit-scale intercept shift such that weighted.mean(x,w) = target
}
\keyword{internal}
