% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-shift.R
\name{normalize_preds}
\alias{normalize_preds}
\title{Normalize predictions to sum to a constraint within rows}
\usage{
normalize_preds(preds, vars, constraints = rep(1, length(vars)))
}
\arguments{
\item{preds}{A $D \times N \times K$ array or an $N \times K$ data frame
storing predictions from the poststratification table to be normalized.
If an array, the first index stores the draw number (e.g. from the posterior),
the second indexes rows in the poststratification table, and the third indexes
columns.}

\item{vars}{A list of sets of variables that must sum to one. See details.}

\item{constraints}{A numeric vector, of the same length as \code{vars}, that lists
what the parameters must sum to. Alternatively, a list of the same length
as \code{vars}, each element of which is a numeric vector that specifies what
each row must sum to. Each element of this list must be the same length as
the number of rows in \code{preds} (i.e. $N$). Defaults to 1.}
}
\description{
Rescales predictions of categorical variables to sum to a particular
constraint. This is useful when the predictions for each category are modeled
as independent --- for example, when three vote choice categories (Dem, Rep,
Other) are modeled in separate equations --- but they must sum to a specific
value.
}
\details{
For each set of variables in \code{vars}, this function rescales the
variables so that their sum equals their corresponding constraint. For
example, if \code{vars = list(c(v1, v2), c(v3, v4))} and \code{constraints = c(1, 2)},
this function will enforce the following constraints for each row in \code{preds}:
\code{v1 + v2 = 1} and \code{v3 + v4 = 2}.
}
