% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-shift.R
\name{calibrate_preds}
\alias{calibrate_preds}
\title{Recalibrate probabilities by applying a user-supplied logit shift}
\usage{
calibrate_preds(
  ps_table,
  shifts,
  preds,
  geography,
  shift_suffix = "shift",
  calib_suffix = "calib"
)
}
\arguments{
\item{ps_table}{A data frame storing poststratification cells with model-based
predictions attached.}

\item{shifts}{A data frame of logit shift parameters.}

\item{preds}{Variables in \code{ps_table} that store model-based predictions
for two outomes and prefix for shift parameters in \code{shifts}}

\item{geography}{Geography variable linking \code{ps_table} and \code{shifts}.}

\item{shift_suffix}{Suffix for the shift variables, e.g. if the first outcome
variable is called \code{presvote} then the shift variable will be called
\code{presvote_shift}.}

\item{calib_suffix}{Suffix for recalibrated vote probabilities.}
}
\description{
Takes the poststratification table and a table of logit shifts and returns
the recalibrated probabilities.
}
