% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_estimates.R
\name{generate_cell_estimates}
\alias{generate_cell_estimates}
\title{Generate estimates for each poststratification cell}
\usage{
generate_cell_estimates(
  model,
  ps_table,
  outcomes = NULL,
  summarize = TRUE,
  draw_ids = NULL,
  outcome_suffix = "",
  se_suffix = "_se",
  control = list(batch_size = 50)
)
}
\arguments{
\item{model}{A \code{brmsfit} object with multivariate binary outcomes.}

\item{ps_table}{A data frame of poststratification cells (must match model variables).}

\item{outcomes}{Optional character vector of outcome variables to calibrate.
If \code{NULL}, the outcome names are inferred from the \code{model} formula.}

\item{summarize}{Should the function return a summary data frame with posterior
means and SDs? If \code{FALSE}, returns an array of draws from the posterior.}

\item{draw_ids}{Optional vector of posterior draw indices to use. If NULL
(the default), use all draws.}

\item{outcome_suffix}{Optional suffix to append to posterior mean columns. Defaults to no suffix.}

\item{se_suffix}{Suffix to append to posterior SD columns. Defaults to "_se".}

\item{control}{A list of control parameters. See details for list of supported
controls.}
}
\value{
If \code{summarize = TRUE}, returns \code{ps_table} with additional columns
for outcome posterior means and posterior standard deviations.
For example, if the outcomes in the \code{brms} model are \code{voteshare} and \code{turnout}
and default suffix arguments are used, the addition columns will be called
\code{voteshare},\code{voteshare_se}, \code{turnout}, and \code{turnout_se}.

If \code{summarize = FALSE}, a 3D array with dimensions
\code{N draws} \eqn{\times} \code{N cells} \eqn{\times} \code{N outcomes}.
}
\description{
Uses a fitted multivariate \code{brms} model to generate predictions for each row
in a poststratification table. Internally, the function uses \link[brms:posterior_epred]{brms::posterior_epred}
to generate draws from the posterior for each cell. There are two return options. If
\code{summarize = TRUE} (the default), the function summarizes the posterior draws
and returns the poststratification
table with added columns for posterior mean and SD of each outcome.
If \code{summarize = FALSE}, the function returns a 3D array that
contains each draw from the posterior for each cell and outcome. In this case,
the function is essentially a wrapper around \link[brms:posterior_epred]{brms::posterior_epred}.
}
\details{
This function generates estimates of the average outcome in each cell of the
poststratification table.

This function is memory and computationally intensive, especially when
working with large poststratification tables. To reduce computational burden when
developing code, you can reduce the number of draws via \code{draw_ids} or work with
a subset of the poststratification table. However, all draws should be used for
final inference.

When \code{summarize = TRUE}, it is possible to reduce memory usage by processing
the posterior samples in batches, which is controlled by the \code{batch_size}
item in the \code{control} list. If you have a large amount of memory available,
increasing the batch size may speed up computation.

When \code{summarize = FALSE}, the function returns a array of posterior draws,
which may have large dimensions in some cases.
}
\examples{
## See vignette for usage and workflow.

}
